requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
class-registry==2.1.2
click==8.0.3
code128==0.2b2
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.1
django-filter==21.1
django-widget-tweaks==1.4.12
filelock==3.4.2
filters==1.3.2
Flask==2.0.2
Flask-Babel==2.0.0
Flask-MySQL==1.5.2
Flask-Table==0.5.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.1
outcome==1.1.0
panda==0.3.1
pandas==1.4.0
pdfkit==1.0.0
Pillow==9.0.1
platformdirs==2.4.1
protobuf==3.19.4
pycodestyle==2.8.0
pycparser==2.21
pymssql==2.2.4
PyMySQL==1.0.2
pyodbc==4.0.32
pyOpenSSL==22.0.0
python-barcode==0.13.1
python-dateutil==2.8.2
pytz==2021.3
regex==2022.3.15
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.0
webdriver-manager==3.5.2
Werkzeug==2.0.2
whitenoise==6.0.0
wsproto==1.0.0
